requirements.txt
albumentations==1.3.0
ffmpeg_python==0.2.0
imageio==2.9.0
imutils==0.5.4
matplotlib==3.7.1
mediapipe==0.10.8
networkx==3.1
numpy==1.23.5
pandas==1.3.4
Pillow==10.1.0
qc_procrustes==1.0.0
scikit_learn==1.2.2
scipy==1.11.4
seaborn==0.11.2
setuptools==61.2.0
simdkalman==1.0.2
sk_video==1.1.10
torch==1.12.1
torchvision==0.13.1
tqdm==4.65.0
